requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
xgboost>=2.1.0
joblib>=1.4.0
boto3>=1.34.0
structlog>=24.2.0
click>=8.1.7
pyyaml>=6.0.2
requests>=2.32.0
faker>=28.0.0
shap>=0.46.0
duckdb>=1.1.0
great-expectations>=0.18.0
evidently>=0.4.33
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
